requirements/base.txt
djangorestframework==3.14.0
drf-yasg==1.21.7
django-cors-headers==4.3.1

psycopg2-binary==2.9.9

environs==10.3.0
gunicorn==21.2.0
pillow==10.2.0

django-silk==5.1.0
django-debug-toolbar==4.3.0

djangorestframework-simplejwt==5.3.1
django-ckeditor==6.7.1
